--- src/Contracts/Traceable.php
@@ -1,16 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace RicorocksDigitalAgency\Soap\Contracts;
-
-interface Traceable
-{
-    public function __getLastRequest(): ?string;
-
-    public function __getLastResponse(): ?string;
-
-    public function __getLastRequestHeaders(): ?string;
-
-    public function __getLastResponseHeaders(): ?string;
-}
