+++ src/Request/SoapPhpClientRequest.php
@@ -6,7 +6,7 @@ namespace RicorocksDigitalAgency\Soap\Request;
 
 use Closure;
 use RicorocksDigitalAgency\Soap\Contracts\Builder;
-use RicorocksDigitalAgency\Soap\Contracts\Client;
+use RicorocksDigitalAgency\Soap\Contracts\PhpSoap\Client;
 use RicorocksDigitalAgency\Soap\Contracts\Request;
 use RicorocksDigitalAgency\Soap\Contracts\Soapable;
 use RicorocksDigitalAgency\Soap\Header;
@@ -14,7 +14,7 @@ use RicorocksDigitalAgency\Soap\Response\Response;
 use RicorocksDigitalAgency\Soap\Support\Tracing\Trace;
 use SoapHeader;
 
-final class SoapClientRequest implements Request
+final class SoapPhpClientRequest implements Request
 {
     private Builder $builder;
