+++ tests/Unit/Headers/HeadersTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-use RicorocksDigitalAgency\Soap\Request\SoapClientRequest;
+use RicorocksDigitalAgency\Soap\Request\SoapPhpClientRequest;
 
 it('can set headers')
     ->fake()
@@ -10,7 +10,7 @@ it('can set headers')
         ->withHeaders($this->soap()->header('Auth', 'test.com')->data(['foo' => 'bar']))
         ->call('Add', ['intA' => 10, 'intB' => 25])
     )
-    ->assertSent(fn (SoapClientRequest $request) => $request->getHeaders() == [
+    ->assertSent(fn (SoapPhpClientRequest $request) => $request->getHeaders() == [
         $this->soap()->header('Auth', 'test.com')->data(['foo' => 'bar']),
     ]);
 
@@ -23,7 +23,7 @@ it('can define multiple headers in the same method')
         )
         ->call('Add', ['intA' => 10, 'intB' => 25])
     )
-    ->assertSent(fn (SoapClientRequest $request) => $request->getHeaders() == [
+    ->assertSent(fn (SoapPhpClientRequest $request) => $request->getHeaders() == [
         $this->soap()->header('Auth', 'test.com')->data(['foo' => 'bar']),
         $this->soap()->header('Brand', 'test.com')->data(['hello' => 'world']),
     ]);
@@ -37,7 +37,7 @@ it('can define multiple headers with an array in the same method')
         ])
         ->call('Add', ['intA' => 10, 'intB' => 25])
     )
-    ->assertSent(fn (SoapClientRequest $request) => $request->getHeaders() == [
+    ->assertSent(fn (SoapPhpClientRequest $request) => $request->getHeaders() == [
         $this->soap()->header('Auth', 'test.com')->data(['foo' => 'bar']),
         $this->soap()->header('Brand', 'test.com')->data(['hello' => 'world']),
     ]);
@@ -51,7 +51,7 @@ it('can define multiple headers using a collection in the same method')
         ]))
         ->call('Add', ['intA' => 10, 'intB' => 25])
     )
-    ->assertSent(fn (SoapClientRequest $request) => $request->getHeaders() == [
+    ->assertSent(fn (SoapPhpClientRequest $request) => $request->getHeaders() == [
         $this->soap()->header('Auth', 'test.com')->data(['foo' => 'bar']),
         $this->soap()->header('Brand', 'test.com')->data(['hello' => 'world']),
     ]);
@@ -63,7 +63,7 @@ it('can define multiple headers in multiple methods')
         ->withHeaders($this->soap()->header('Brand', 'test.com')->data(['hello' => 'world']))
         ->call('Add', ['intA' => 10, 'intB' => 25])
     )
-    ->assertSent(fn (SoapClientRequest $request) => $request->getHeaders() == [
+    ->assertSent(fn (SoapPhpClientRequest $request) => $request->getHeaders() == [
         $this->soap()->header('Auth', 'test.com')->data(['foo' => 'bar']),
         $this->soap()->header('Brand', 'test.com')->data(['hello' => 'world']),
     ]);
@@ -80,7 +80,7 @@ it('can create a header without any parameters and be composed fluently')
         )
         ->call('Add', ['intA' => 10, 'intB' => 25])
     )
-    ->assertSent(fn (SoapClientRequest $request) => $request->getHeaders() == [
+    ->assertSent(fn (SoapPhpClientRequest $request) => $request->getHeaders() == [
         $this->soap()->header('Auth', 'test.com')->data(['foo' => 'bar'])->mustUnderstand()->actor('this.test'),
     ]);
 
@@ -90,7 +90,7 @@ it('can set up a header using a SoapVar')
         ->withHeaders($this->soap()->header('Auth', 'test.com', new SoapVar(['foo' => 'bar'], null)))
         ->call('Add', ['intA' => 10, 'intB' => 25])
     )
-    ->assertSent(fn (SoapClientRequest $request) => $request->getHeaders() == [
+    ->assertSent(fn (SoapPhpClientRequest $request) => $request->getHeaders() == [
         $this->soap()->header('Auth', 'test.com')->data(new SoapVar(['foo' => 'bar'], null)),
     ]);
 
@@ -101,7 +101,7 @@ it('does not require the data parameter')
         ->withHeaders($this->soap()->header('Brand', 'test.com', null))
         ->call('Add', ['intA' => 10, 'intB' => 25])
     )
-    ->assertSent(fn (SoapClientRequest $request) => $request->getHeaders() == [
+    ->assertSent(fn (SoapPhpClientRequest $request) => $request->getHeaders() == [
         $this->soap()->header('Auth', 'test.com', null),
         $this->soap()->header('Brand', 'test.com', null),
     ]);
