+++ tests/Unit/Options/AuthTest.php
@@ -2,14 +2,14 @@
 
 declare(strict_types=1);
 
-use RicorocksDigitalAgency\Soap\Request\SoapClientRequest;
+use RicorocksDigitalAgency\Soap\Request\SoapPhpClientRequest;
 
 it('can use basic auth')
     ->fake()
     ->soap()->to(EXAMPLE_SOAP_ENDPOINT)
     ->withBasicAuth('hello', 'world')
     ->call('Add', ['intA' => 10, 'intB' => 25])
-    ->test()->assertSent(function (SoapClientRequest $request) {
+    ->test()->assertSent(function (SoapPhpClientRequest $request) {
         return $request->getOptions() == [
                 'authentication' => SOAP_AUTHENTICATION_BASIC,
                 'login' => 'hello',
@@ -22,7 +22,7 @@ it('can use digest auth')
     ->soap()->to(EXAMPLE_SOAP_ENDPOINT)
     ->withDigestAuth('hello', 'world')
     ->call('Add', ['intA' => 10, 'intB' => 25])
-    ->test()->assertSent(function (SoapClientRequest $request) {
+    ->test()->assertSent(function (SoapPhpClientRequest $request) {
         return $request->getOptions() == [
                 'authentication' => SOAP_AUTHENTICATION_DIGEST,
                 'login' => 'hello',
